move test setup logic to helper module
